+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ### Added
+- ** Add Hitide Generate Configuration **
+  - [issue/26](https://github.com/podaac/tig/issues/26):
 ### Changed
 ### Deprecated
 ### Removed